docker-post/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
idna==3.3
requests==2.26.0
six==1.15.0
urllib3==1.26.7
wykop-sdk-reborn==0.9.1
